Добавлено использвоание параметров длины и массы из UnitsNet в преобразовании в DTO/Model, значения внутри системы - граммы и миллиметры

+++ TaskControl.InformationModule/DataAccess/Model/ItemModel.cs
@@ -11,10 +11,10 @@ namespace TaskControl.InformationModule.DataAccess.Model
     public class ItemModel
     {
         [Column("item_id")][PrimaryKey][Identity] public int ItemId { get; set; }
-        [Column("weight")][NotNull] public decimal Weight { get; set; }
-        [Column("length")][NotNull] public decimal Length { get; set; }
-        [Column("width")][NotNull] public decimal Width { get; set; }
-        [Column("height")][NotNull] public decimal Height { get; set; }
+        [Column("weight")][NotNull] public double Weight { get; set; }
+        [Column("length")][NotNull] public double Length { get; set; }
+        [Column("width")][NotNull] public double Width { get; set; }
+        [Column("height")][NotNull] public double Height { get; set; }
         [Column("created_at")][NotNull] public DateTime CreatedAt { get; set; }
     }
 }
